requirements.txt
kafka-python
fastavro
requests
mcp-server
trino
python-dotenv
